refactor(core): consolidate allTeams/updateTeam/findTeam/isTeamAlive helpers into phase-helpers
Eliminates duplicate implementations of four utility functions that were
copied independently across phase-helpers.ts, session.ts, and
strategy/strategy.ts:

- allTeams, findTeam, isTeamAlive, updateTeam are now exported from the
  single source of truth (phase-helpers.ts)
- session.ts imports them from phase-helpers.ts instead of defining local
  copies
- strategy/strategy.ts re-exports findTeam and isTeamAlive from
  phase-helpers.ts to preserve the public package API, removing the
  private allTeams duplicate

Resolves the code-review finding on the mixed-control session manager PR.
