+++ packages/core/src/strategy/strategy.ts
@@ -1,7 +1,6 @@
 import type { BattlePlay } from '../logic/battle.ts';
 import type { RevivalAction, RoundState } from '../logic/round.ts';
 import type { MovementChoice } from '../logic/runner.ts';
-import { ELEMENT_AXIS, type TeamState } from '../types/grid.ts';
 import type { TeamId } from '../types/token.ts';
 
 type ExplicitStrategyMovementChoice = Omit<
@@ -36,25 +35,5 @@ export type TeamStrategy = {
   ) => RevivalAction | null;
 };
 
-function allTeams(state: RoundState): TeamState[] {
-  const teams: TeamState[] = [];
-  for (const x of ELEMENT_AXIS) {
-    for (const y of ELEMENT_AXIS) {
-      teams.push(...state.grid[x][y]);
-    }
-  }
-  return teams;
-}
-
 /** Returns the current team snapshot for a team id, if still on the board. */
-export function findTeam(
-  state: RoundState,
-  teamId: TeamId,
-): TeamState | undefined {
-  return allTeams(state).find((team) => team.teamNumber === teamId);
-}
-
-/** Returns true when the team still has at least one living player. */
-export function isTeamAlive(team: TeamState): boolean {
-  return team.players.some((player) => player.life > 0);
-}
+export { findTeam, isTeamAlive } from '../logic/phase-helpers.ts';
